+++ apps/next-client/src/components/LocaleSwitcher.tsx
@@ -1,18 +1,46 @@
+'use client';
+
+import { useTransition } from 'react';
+import { useParams } from 'next/navigation';
 import { useLocale, useTranslations } from 'next-intl';
-import { routing } from '@/i18n/routing';
-import LocaleSwitcherSelect from './LocaleSwitcherSelect';
+import Select, { SelectChangeEvent } from '@mui/material/Select';
+import MenuItem from '@mui/material/MenuItem';
+import { routing, usePathname, useRouter } from '@/i18n/routing';
+import { Locales } from '@/types';
 
 export default function LocaleSwitcher() {
+  const router = useRouter();
+  const currentLocale = useLocale();
+  const [isPending, startTransition] = useTransition();
+  const pathname = usePathname();
+  const params = useParams();
   const t = useTranslations('LocaleSwitcher');
-  const locale = useLocale();
+
+  function onSelectChange(event: SelectChangeEvent) {
+    const nextLocale = event.target.value as Locales;
+
+    startTransition(() => {
+      router.replace(
+        // @ts-expect-error -- TypeScript will validate that only known `params`
+        // are used in combination with a given `pathname`. Since the two will
+        // always match for the current route, we can skip runtime checks.
+        { pathname, params },
+        { locale: nextLocale }
+      );
+    });
+  }
 
   return (
-    <LocaleSwitcherSelect defaultValue={locale} label={t('label')}>
+    <Select
+      value={currentLocale}
+      disabled={isPending}
+      onChange={onSelectChange}
+    >
       {routing.locales.map(cur => (
-        <option key={cur} value={cur}>
+        <MenuItem key={cur} value={cur}>
           {t('locale', { locale: cur })}
-        </option>
+        </MenuItem>
       ))}
-    </LocaleSwitcherSelect>
+    </Select>
   );
 }
