+++ apps/next-client/src/components/index.ts
@@ -2,5 +2,4 @@ export { default as AppBar } from './AppBar';
 export { default as BaseLayout } from './BaseLayout';
 export * from './Link';
 export { default as LocaleSwitcher } from './LocaleSwitcher';
-export { default as LocaleSwitcherSelect } from './LocaleSwitcherSelect';
 export { default as NotFoundPage } from './NotFoundPage';
